chore: refactor settings handling into new settings service
