+++ Chefs/Services/Sharing/IShareService.cs
@@ -5,5 +5,13 @@ namespace Chefs.Services.Sharing;
 /// </summary>
 public interface IShareService
 {
+	///<summary>
+	/// Open native sharing for a recipe and its steps
+	/// </summary>
+	/// <param name="recipe">recipe to share</param>
+	/// <param name="steps">recipe's steps</param>
+	/// <param name="ct"></param>
+	/// <returns>
+	/// </returns>
 	Task ShareRecipe(Recipe recipe, IImmutableList<Step> steps, CancellationToken ct);	
 }
