+++ Chefs/Services/Users/IUserService.cs
@@ -28,15 +28,6 @@ public interface IUserService
 	/// </returns>
 	ValueTask Update(User user, CancellationToken ct);
 
-	///<summary>
-	/// Gets chef settings
-	/// </summary>
-	/// <param name="ct"></param>
-	/// <returns>
-	/// App settings from the phone
-	/// </returns>
-	ValueTask<AppConfig> GetSettings(CancellationToken ct);
-
 	/// <summary>
 	/// Porpular creators related with the recipes
 	/// </summary>
@@ -54,27 +45,6 @@ public interface IUserService
 	/// User
 	/// </returns>
 	ValueTask<User> GetById(Guid userId, CancellationToken ct);
-
-	///<summary>
-	/// Update app settings
-	/// </summary>
-	/// <param name="chefSettings">new settings</param>
-	/// <param name="ct"></param>
-	/// <returns>
-	/// </returns>
-	Task SetSettings(AppConfig chefSettings, CancellationToken ct);
-
-	///<summary>
-	/// Update app settings with specified properties without overwriting the rest
-	/// </summary>
-	/// <param name="ct"></param>
-	/// <param name="title">App title</param>
-	/// <param name="isDark">App theme flag</param>
-	/// <param name="notification">User notifications flag</param>
-	/// <param name="accentColor">Accent color</param>
-	/// <returns>
-	/// </returns>
-	Task UpdateSettings(CancellationToken ct, string? title = null, bool? isDark = null, bool? notification = null, string? accentColor = null);
 
 	// <summary>
 	// Authentication method
